refactor(trace-model): refactor trace model package
