+++ metadata-io/src/main/java/com/linkedin/metadata/trace/KafkaTraceReader.java
@@ -6,10 +6,10 @@ import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.linkedin.common.urn.Urn;
 import com.linkedin.data.template.RecordTemplate;
-import com.linkedin.metadata.systemmetadata.TraceStorageStatus;
-import com.linkedin.metadata.systemmetadata.TraceWriteStatus;
 import com.linkedin.mxe.SystemMetadata;
 import com.linkedin.util.Pair;
+import io.datahubproject.openapi.v1.models.TraceStorageStatus;
+import io.datahubproject.openapi.v1.models.TraceWriteStatus;
 import java.time.Duration;
 import java.util.Collection;
 import java.util.Collections;
