+++ metadata-io/src/main/java/com/linkedin/metadata/trace/TraceServiceImpl.java
@@ -11,9 +11,6 @@ import com.linkedin.metadata.models.registry.EntityRegistry;
 import com.linkedin.metadata.run.AspectRowSummary;
 import com.linkedin.metadata.systemmetadata.SystemMetadataService;
 import com.linkedin.metadata.systemmetadata.TraceService;
-import com.linkedin.metadata.systemmetadata.TraceStatus;
-import com.linkedin.metadata.systemmetadata.TraceStorageStatus;
-import com.linkedin.metadata.systemmetadata.TraceWriteStatus;
 import com.linkedin.mxe.FailedMetadataChangeProposal;
 import com.linkedin.mxe.SystemMetadata;
 import com.linkedin.util.Pair;
@@ -21,6 +18,9 @@ import io.datahubproject.metadata.context.OperationContext;
 import io.datahubproject.metadata.context.TraceContext;
 import io.datahubproject.metadata.context.TraceIdGenerator;
 import io.datahubproject.metadata.exception.TraceException;
+import io.datahubproject.openapi.v1.models.TraceStatus;
+import io.datahubproject.openapi.v1.models.TraceStorageStatus;
+import io.datahubproject.openapi.v1.models.TraceWriteStatus;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
