+++ metadata-io/src/test/java/com/linkedin/metadata/trace/BaseKafkaTraceReaderTest.java
@@ -12,11 +12,11 @@ import com.linkedin.common.urn.Urn;
 import com.linkedin.common.urn.UrnUtils;
 import com.linkedin.data.template.RecordTemplate;
 import com.linkedin.data.template.StringMap;
-import com.linkedin.metadata.systemmetadata.TraceStorageStatus;
-import com.linkedin.metadata.systemmetadata.TraceWriteStatus;
 import com.linkedin.mxe.SystemMetadata;
 import com.linkedin.util.Pair;
 import io.datahubproject.metadata.context.TraceContext;
+import io.datahubproject.openapi.v1.models.TraceStorageStatus;
+import io.datahubproject.openapi.v1.models.TraceWriteStatus;
 import java.io.IOException;
 import java.time.Duration;
 import java.util.Collections;
