+++ metadata-io/src/test/java/com/linkedin/metadata/trace/TraceServiceImplTest.java
@@ -25,9 +25,6 @@ import com.linkedin.metadata.EventUtils;
 import com.linkedin.metadata.entity.EntityService;
 import com.linkedin.metadata.run.AspectRowSummary;
 import com.linkedin.metadata.systemmetadata.SystemMetadataService;
-import com.linkedin.metadata.systemmetadata.TraceStatus;
-import com.linkedin.metadata.systemmetadata.TraceStorageStatus;
-import com.linkedin.metadata.systemmetadata.TraceWriteStatus;
 import com.linkedin.mxe.FailedMetadataChangeProposal;
 import com.linkedin.mxe.MetadataChangeProposal;
 import com.linkedin.mxe.SystemMetadata;
@@ -35,6 +32,9 @@ import com.linkedin.util.Pair;
 import io.datahubproject.metadata.context.OperationContext;
 import io.datahubproject.metadata.context.TraceContext;
 import io.datahubproject.metadata.context.TraceIdGenerator;
+import io.datahubproject.openapi.v1.models.TraceStatus;
+import io.datahubproject.openapi.v1.models.TraceStorageStatus;
+import io.datahubproject.openapi.v1.models.TraceWriteStatus;
 import io.datahubproject.test.metadata.context.TestOperationContexts;
 import java.time.Instant;
 import java.util.Collections;
