+++ metadata-service/openapi-servlet/models/src/main/java/io/datahubproject/openapi/v1/models/TraceStorageStatus.java
@@ -1,4 +1,4 @@
-package com.linkedin.metadata.systemmetadata;
+package io.datahubproject.openapi.v1.models;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import io.datahubproject.metadata.exception.TraceException;
