+++ metadata-service/openapi-servlet/models/src/main/java/io/datahubproject/openapi/v1/models/TraceWriteStatus.java
@@ -1,4 +1,4 @@
-package com.linkedin.metadata.systemmetadata;
+package io.datahubproject.openapi.v1.models;
 
 public enum TraceWriteStatus {
   // error occurred during processing
