+++ metadata-service/openapi-servlet/src/test/java/io/datahubproject/openapi/operations/v1/TraceControllerTest.java
@@ -18,15 +18,15 @@ import com.linkedin.common.urn.Urn;
 import com.linkedin.common.urn.UrnUtils;
 import com.linkedin.metadata.models.registry.EntityRegistry;
 import com.linkedin.metadata.systemmetadata.TraceService;
-import com.linkedin.metadata.systemmetadata.TraceStatus;
-import com.linkedin.metadata.systemmetadata.TraceStorageStatus;
-import com.linkedin.metadata.systemmetadata.TraceWriteStatus;
 import io.datahubproject.metadata.context.ObjectMapperContext;
 import io.datahubproject.metadata.context.OperationContext;
 import io.datahubproject.metadata.context.TraceContext;
 import io.datahubproject.openapi.config.TracingInterceptor;
 import io.datahubproject.openapi.v1.models.TraceRequestV1;
 import io.datahubproject.openapi.v1.models.TraceResponseV1;
+import io.datahubproject.openapi.v1.models.TraceStatus;
+import io.datahubproject.openapi.v1.models.TraceStorageStatus;
+import io.datahubproject.openapi.v1.models.TraceWriteStatus;
 import io.datahubproject.test.metadata.context.TestOperationContexts;
 import java.util.List;
 import java.util.Map;
